Add files via upload
Script to filter metagenomic OTU count data with a double-threshold approach, so that OTUs present in many samples and at the same time are identified. This curated dataset is intended to be used by machine learning tools, neural networks, and other geolocation prediction approaches according to bacterial soil taxonomic abundances.