Update variable_selection.R
Update does not depend on DESeq2 nor fdrtools, instead it depends on purrr and broom. 
With this modification, it works for all hierarchical levels (separated by kingdoms) and for assembly data.
It also implements a negative binomial model for each OTU and each pair of (city x year) instead of the full model. 
TODO: comment the code and implement the zero inflated negative binomial model (an automatic model selection could be implemented)